distributions: split mod.rs into based on distribution type
